Added dev directory for code written by me while I work through this course. The readme in the dev directory describes the intent.
"section-2" - directory for any work done during Section 2 of the
course.

I intend to use a directory for each section of the course and nested
directories for the projects in the section.

I could see "section-2" directory getting cluttered quickly.

There is a 100 days challenge aspect of the course. Each video has the
day you would need to view it in order to complete the course in 100
days.

I plan to do it in less than 100 days but I may use that as a way to organize the
course work.