Was able to get the bar at the left side of the `li`. Realized a `border-left` would do just that. I think it works out well.
In regard to my css selector choice I usually like to use combinators
but thinking about it now maybe that's not too good of an idea because
then if I edit the html structure I also have to edit the CSS
selectors.

Not sure what the rule of thumb is but also considering it could be
something you account for when you have to deal with it. I think in this
situation where it's a small excercise with no other pages
or re-use of CSS going on, using combinators relient on HTML structure
will suffice.

The rest I think was pretty easy. I didn't watch the video yet. This is
my version before making changes from watching the video. I may be off a few pixels or maybe he
had a different way of doing it but I think it turned out well.