Learned about `form` semantics and how to structure the HTML so it is easier to maintain. This involved the use of `section` tags to group fields together. Using `ul` and `li` tags to structure `label` and `input` tags together.
Learned about validation attributes like `required` `min` `minlength`.

Learned about the use of the `novalidation` attribute on the `form`
element. Of note, is the fact that the `novalidation` attribute is a
`boolean` attribute and does not require a value. A `boolean`
attribute's presences on a tag denotes it is active.

Next up is a challenge where I will create a form on my own.