Working through section-12: Advanced JavaScript: Working with the DOM.
Learned about how to use the `document` object and some methods on it to
interact with the DOM.

Learned how to drill into properties on the `document` object to access
HTML elements. Also learned about methods on the `document` object like
`querySelector`, `getElementById`, and `getElementByClass`.

Next is on to Event Handling.