Finished Section 21: Exploring More Advanced JavaScript Features
Learned about default parameters on functions and how that turns a
parameter into an option parameter.

Learned about Rest Parameters & the spread operator.
Rest parameters - used for function definitions and allows for a
variable amount of paramters to be passed to a function.
Spread operator - allows coder to pull out values of an array or object
when a list of key value pairs or values of an array when a list
values/key value pairs are needed

Learned about template literals.
Allows us to construct strings with placeholders which is easier to read
and more convenient than concatenating strings with the plus operator.

Learned the difference between primitive & Reference values.
Objects are reference values which is stored on the heap and stored in
the variable is a pointer (the adress) to the value in memory.
You can add values to a constant object or array using push method or
the . notation - because that does not change the constant which
contains the address to the value (which didnt change).

Learned can use spread operator to create copies of objects/arrays so
you do not have unwanted mutations to the original array/object.

Learned about error handling and using `try/catch`.

Learned about classes - allowing coder to create a blue print for the
object using the constructor for the class.

Learned about the `this` keyword which refers to an object instance of
the class.

The `new` keyword allows coder to create new instances of the class by
excuting the class and using the `new` keyword.

Learned about Async code.
Some operations can take longer and may want to execute it
asynchronously - either using call back or promise.

If using promises, you can turn the overall function that wraps the code
into an async fuction then use the `await` keyword to go back to more
synchronous looking code. Even though it's transformed under the hood to
promises and `then` functions.