Finished Sections 22 - 24.
Sections 22 and 23 were really just introductions to the core concepts
of SQL Databases.
All work was done using a local instance of MySQL and the MySQL
workbench tool.

Section 24 was about connecting to a MySQL database in a Node.js Express
application.

We implemented basic CRUD operations.

I learned about patterns for creating routes as well.