Finished implementing product creation and product updating. Implemented deletion of product image when changing the product image. Need to implement deletion of prod image when deleting product next.
I haven't added any checks when deleting a product.
For example, if there exist open orders with the product trying to be deleted, it probaly should not be allowed
to get deleted. But that's introducing some constraints I am not sure I want to do just yet.

After that it'll be order maintenance.

That needs to be implemented for the customer and admin.

Not sure how in depth I will go there. I am going to review the module introduction video again
to see how in-depth order maintenance is for customers & for administrators.